+++ pytao/constraints/observables/twiss.py
@@ -3,7 +3,7 @@ from typing import Annotated, Literal, Union
 
 from pydantic import BaseModel, Field
 
-from pytao.check.observables.base import CheckResult
+from pytao.constraints.observables.base import CheckResult
 
 
 class TwissComparisonMethod(BaseModel, ABC):
